some edits

+++ qr3/flops/flops_lapack_geqr2.c
@@ -7,11 +7,10 @@ long int flops_lapack_geqr2( int int_m, int int_n ){
 	m = ( long int ) int_m;
 	n = ( long int ) int_n;
 
-	flops = ( (( long int ) 6) * m * n * n 
-	- (( long int ) 2) * n * n * n 
-	+ (( long int ) 3) * m * n 
-	+ (( long int ) 17) * n )
-	/ (( long int ) 3);
+	flops = ( 6 * m * n * n 
+	       -  2 * n * n * n 
+	       +  3 * m * n 
+	       + 17 * n ) /  3;
 
 	return flops;
